requirements.txt
aider==0.2.4
aider-chat==0.48.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.33.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
entrypoints==0.4
Faker==26.3.0
favicon==0.7.0
filelock==3.15.4
flake8==7.1.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
grep-ast==0.3.3
groq==0.9.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
litellm==1.42.9
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.4.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
openai==1.37.2
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
plotly==5.23.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
pyarrow==17.0.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pymdown-extensions==10.9
pypandoc==1.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.37.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.6
streamlit-faker==0.0.3
streamlit-float==0.3.5
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-option-menu==0.3.13
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.33.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.2
anthropic
